Rip out the previous ReplicaSync fix attempt.
