+++ server/datastore/mysql/mysql.go
@@ -106,17 +106,6 @@ type Datastore struct {
 	testBatchSetMDMWindowsProfilesErr string
 }
 
-// MasterStatus is a struct that holds the file and position of the master, retrieved by SHOW MASTER STATUS
-type MasterStatus struct {
-	File     string
-	Position uint64
-}
-
-// ReplicaStatus is a struct that holds the status information of the read replica, retrieved by SHOW SLAVE STATUS
-type ReplicaStatus struct {
-	ExecSourceLogPosition uint64
-}
-
 // reader returns the DB instance to use for read-only statements, which is the
 // replica unless the primary has been explicitly required via
 // ctxdb.RequirePrimary.
@@ -1361,155 +1350,4 @@ func batchProcessDB[T any](
 		}
 	}
 	return nil
-}
-
-func (ds *Datastore) ReplicaSync(ctx context.Context) error {
-	if ds.primary == ds.replica {
-		// No replica present, nothing to sync
-		return nil
-	}
-	// Grab the current primary's binlog position
-	ms, err := ds.MasterStatus(ctx)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Master status: %+v\n", ms)
-	// We read replica status right away because replica may already be in sync if there were no recent writes
-	rs, err := ds.ReplicaStatus(ctx)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Replica status: %+v\n", rs)
-	if rs.ExecSourceLogPosition < ms.Position {
-		// Wait for replica to catch up. In production environment, replication time is usually under 30 milliseconds.
-		done := make(chan error, 1)
-		go func() {
-			for rs.ExecSourceLogPosition < ms.Position {
-				time.Sleep(30 * time.Millisecond)
-				rs, err = ds.ReplicaStatus(ctx)
-				if err != nil {
-					done <- err
-					return
-				}
-				fmt.Printf("Replica status: %+v\n", rs)
-			}
-			done <- nil
-		}()
-		select {
-		case err := <-done:
-			return err
-		case <-time.After(60 * time.Second): // BOZO
-			return ctxerr.New(ctx, "replica sync timeout: replica did not catch up to the master in 5 seconds")
-		}
-	}
-	return nil
-}
-
-func (ds *Datastore) MasterStatus(ctx context.Context) (MasterStatus, error) {
-
-	rows, err := ds.writer(ctx).Query("SHOW MASTER STATUS")
-	if err != nil {
-		return MasterStatus{}, ctxerr.Wrap(ctx, err, "show master status")
-	}
-	defer rows.Close()
-
-	// Since we don't control the column names, and we want to be future compatible,
-	// we only scan for the columns we care about.
-	ms := MasterStatus{}
-	// Get the column names from the query
-	columns, err := rows.Columns()
-	if err != nil {
-		return ms, ctxerr.Wrap(ctx, err, "get columns")
-	}
-	numberOfColumns := len(columns)
-	for rows.Next() {
-		cols := make([]interface{}, numberOfColumns)
-		for i := range cols {
-			cols[i] = new(string)
-		}
-		err := rows.Scan(cols...)
-		if err != nil {
-			return ms, ctxerr.Wrap(ctx, err, "scan row")
-		}
-		for i, col := range cols {
-			switch columns[i] {
-			case "File":
-				ms.File = *col.(*string)
-			case "Position":
-				ms.Position, err = strconv.ParseUint(*col.(*string), 10, 64)
-				if err != nil {
-					return ms, ctxerr.Wrap(ctx, err, "parse Position")
-				}
-
-			}
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return ms, ctxerr.Wrap(ctx, err, "rows error")
-	}
-	if ms.File == "" || ms.Position == 0 {
-		return ms, ctxerr.New(ctx, "missing required fields in master status")
-	}
-	return ms, nil
-}
-
-func (ds *Datastore) ReplicaStatus(ctx context.Context) (ReplicaStatus, error) {
-	replica := ds.replica.(*sqlx.DB)
-
-	// The 'SHOW SLAVE STATUS' command has been deprecated and replaced by 'SHOW REPLICA STATUS' in MySQL 8.0.22
-	rows, err := replica.Query("SHOW SLAVE STATUS")
-	if err != nil {
-		return ReplicaStatus{}, ctxerr.Wrap(ctx, err, "show replica status")
-	}
-	defer rows.Close()
-
-	// Since we don't control the column names, and we want to be future compatible,
-	// we only scan for the columns we care about.
-	status := ReplicaStatus{}
-	// Get the column names from the query
-	columns, err := rows.Columns()
-	if err != nil {
-		return status, ctxerr.Wrap(ctx, err, "get columns")
-	}
-	numberOfColumns := len(columns)
-	for rows.Next() {
-		cols := make([]interface{}, numberOfColumns)
-		for i := range cols {
-			cols[i] = new(sql.NullString)
-		}
-		err := rows.Scan(cols...)
-		if err != nil {
-			return status, ctxerr.Wrap(ctx, err, "scan row")
-		}
-		for i, col := range cols {
-			value := col.(*sql.NullString)
-			valueString := value.String
-			if !value.Valid {
-				valueString = "INVALID"
-			}
-
-			switch columns[i] {
-			case "Slave_SQL_Running_State", "Replica_SQL_Running_State":
-				fmt.Printf("Column: %s, Value: %s\n", columns[i], valueString)
-			case "Read_Master_Log_Pos", "Read_Source_Log_Pos":
-				fmt.Printf("Column: %s, Value: %s\n", columns[i], valueString)
-			case "Exec_Master_Log_Pos", // MySQL 5.7 name
-				"Exec_Source_Log_Pos": // MySQL 8.0/8.4 name
-				pos := col.(*sql.NullString)
-				status.ExecSourceLogPosition, err = strconv.ParseUint(pos.String, 10, 64)
-				if err != nil {
-					return status, ctxerr.Wrap(ctx, err, "parse Exec_Source_Log_Pos")
-				}
-				// break
-				fmt.Printf("Column: %s, Value: %s\n", columns[i], valueString)
-			}
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return status, ctxerr.Wrap(ctx, err, "rows error")
-	}
-	if status.ExecSourceLogPosition == 0 {
-		return status, ctxerr.New(ctx, "missing required fields in replica status")
-	}
-	return status, nil
 }
